Improve multiselect select-all validation and tests
