+++ e2e_playwright/st_multiselect_test.py
@@ -37,7 +37,7 @@ from e2e_playwright.shared.app_utils import (
     open_popover,
 )
 
-MULTISELECT_COUNT = 33
+MULTISELECT_COUNT = 38
 
 
 def _get_multiselect_input(locator: Locator | Page, label: str) -> Locator:
@@ -69,6 +69,11 @@ def remove_from_multiselect(page: Page, label: str, option_text: str) -> None:
     ms = get_multiselect(page, label)
     ms.get_by_role("button", name=f"Remove {option_text}", exact=True).click()
     wait_for_app_run(page)
+
+
+def _close_dropdown(app: Page) -> None:
+    app.keyboard.press("Escape")
+    expect(app.get_by_role("option")).to_have_count(0)
 
 
 def test_multiselect_on_load(themed_app: Page, assert_snapshot: ImageCompareFunction):
@@ -730,6 +735,110 @@ def test_multiselect_filter_mode_none_disables_typing_but_keeps_selection(app: P
     app.get_by_role("option", name="No", exact=True).click()
     wait_for_app_run(app)
     expect_text(app, "value 23: ['No']")
+
+
+def test_select_all_parameter(app: Page):
+    """select_all controls bulk-action visibility and thresholds."""
+    # False: no bulk action; the first row is the first match.
+    ms_false = get_multiselect(app, "select_all False")
+    ms_false.scroll_into_view_if_needed()
+    input_false = ms_false.locator("input")
+    input_false.click()
+
+    options = app.get_by_role("option")
+    expect(options).to_have_count(3)
+    expect(options.nth(0)).to_have_text("apple")
+    expect(app.get_by_role("option", name="Select all")).not_to_be_visible()
+
+    input_false.press_sequentially("ap")
+    expect(app.get_by_role("option", name="Select 2 matches")).not_to_be_visible()
+    expect(app.get_by_role("option")).to_have_count(2)
+    expect(app.get_by_role("option").nth(0)).to_have_text("apple")
+
+    input_false.press("Enter")
+    wait_for_app_run(app)
+
+    expect_text(app, "select_all False: ['apple']")
+    expect(ms_false.locator('span[title="apricot"]')).not_to_be_visible()
+    _close_dropdown(app)
+
+    # True: Select all is shown.
+    ms_true = get_multiselect(app, "select_all True")
+    ms_true.scroll_into_view_if_needed()
+    ms_true.locator("input").click()
+    expect(app.get_by_role("option", name="Select all")).to_be_visible()
+    options_true = app.get_by_role("option")
+    expect(options_true.nth(0)).to_have_attribute("aria-posinset", "1")
+    first_bg = (
+        options_true.nth(0)
+        .locator("[data-item-hl]")
+        .evaluate("el => getComputedStyle(el).backgroundColor")
+    )
+    second_bg = (
+        options_true.nth(1)
+        .locator("[data-item-hl]")
+        .evaluate("el => getComputedStyle(el).backgroundColor")
+    )
+    assert first_bg != second_bg
+    _close_dropdown(app)
+
+    # Integer threshold uses the filtered selectable count.
+    ms_threshold = get_multiselect(app, "select_all threshold")
+    ms_threshold.scroll_into_view_if_needed()
+    input_threshold = ms_threshold.locator("input")
+    input_threshold.click()
+
+    expect(app.get_by_role("option", name="Select all")).not_to_be_visible()
+    expect(app.get_by_role("option", name="alpha", exact=True)).to_be_visible()
+
+    input_threshold.press_sequentially("al")
+    select_matches = app.get_by_role("option", name="Select 3 matches")
+    expect(select_matches).to_be_visible()
+    expect(app.get_by_role("option", name="Select all")).not_to_be_visible()
+
+    select_matches.click()
+    wait_for_app_run(app)
+
+    expect_text(app, "select_all threshold: ['alpha', 'alpine', 'alta']")
+    _close_dropdown(app)
+
+    # max_selections hides Select all once the cap is reached.
+    ms_max = get_multiselect(app, "select_all with max_selections")
+    ms_max.scroll_into_view_if_needed()
+    input_max = ms_max.locator("input")
+    input_max.click()
+
+    expect(app.get_by_role("option", name="Select all")).to_be_visible()
+    app.get_by_role("option", name="Select all").click()
+    wait_for_app_run(app)
+
+    expect_text(app, "select_all with max_selections: ['red', 'green']")
+    _close_dropdown(app)
+
+    input_max.click()
+    expect(app.get_by_role("option", name="Select all")).not_to_be_visible()
+    expect(app.get_by_test_id("stMultiSelectDropdown")).to_have_text(
+        "You can only select up to 2 options. Remove an option first.",
+        use_inner_text=True,
+    )
+    _close_dropdown(app)
+
+    # Custom chips do not count toward the threshold.
+    ms_chips = get_multiselect(app, "select_all custom chips")
+    ms_chips.scroll_into_view_if_needed()
+    input_chips = ms_chips.locator("input")
+    input_chips.click()
+
+    expect(app.get_by_role("option", name="Select all")).not_to_be_visible()
+    input_chips.press_sequentially("custom")
+    input_chips.press("Enter")
+    wait_for_app_run(app)
+    expect_text(app, "select_all custom chips: ['custom']")
+    _close_dropdown(app)
+
+    input_chips.click()
+    expect(app.get_by_role("option", name="Select all")).not_to_be_visible()
+    expect(app.get_by_role("option", name="one", exact=True)).to_be_visible()
 
 
 # --- Query parameter binding tests ---
